Refactor: UserService 인터페이스 생성
- Proxy class 생성할 때 두번째 인자로 인터페이스가 들어가야 해서 변경
